feat: add caching for notes query

+++ src/services/postgres/NotesService.js
@@ -8,9 +8,10 @@ const AuthorizationError = require('../../exceptions/AuthorizationError');
 const { mapDBToModel } = require('../../utils');
 
 class NoteService {
-  constructor(collaborationService) {
+  constructor(collaborationService, cacheService) {
     this._pool = new Pool();
     this._collaborationService = collaborationService;
+    this._cacheService = cacheService;
   }
 
   /**
@@ -40,6 +41,7 @@ class NoteService {
       throw new InvariantError('Catatan gagal ditambahkan');
     }
 
+    await this._cacheService.delete(`notes:${owner}`);
     return result.rows[0].id;
   }
 
@@ -49,13 +51,24 @@ class NoteService {
    * @returns {Array} An array of all notes in the postgres db.
    */
   async getNotes(owner) {
-    const query = {
-      text: 'SELECT notes.* FROM notes LEFT JOIN collaborations ON collaborations.note_id = notes.id WHERE notes.owner = $1 OR collaborations.user_id = $1 GROUP BY notes.id',
-      values: [owner],
-    };
-    
-    const result = await this._pool.query(query);
-    return result.rows.map(mapDBToModel);
+    const key = `notes:${owner}`;
+
+    try {
+      const result = await this._cacheService.get(key);
+      return JSON.parse(result);
+    } catch {
+      const query = {
+        text: 'SELECT notes.* FROM notes LEFT JOIN collaborations ON collaborations.note_id = notes.id WHERE notes.owner = $1 OR collaborations.user_id = $1 GROUP BY notes.id',
+        values: [owner],
+      };
+
+      const result = await this._pool.query(query);
+      const mappedResult = result.rows.map(mapDBToModel);
+
+      this._cacheService.set(key, JSON.stringify(mappedResult));
+
+      return mappedResult;
+    }
   }
 
   /**
@@ -93,10 +106,10 @@ class NoteService {
    * @throws {Error} If the note with the specified ID is not found.
    * @return {string} The ID of the edited note.
    */
-  async editNoteById(id, { title, body, tags }) {
+  async editNoteById(id, { title, body, tags,  }) {
     const updatedAt = new Date().toISOString();
     const query = {
-      text: 'UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id',
+      text: 'UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id, owner',
       values: [title, body, tags, updatedAt, id],
     };
 
@@ -106,7 +119,10 @@ class NoteService {
       throw new NotFoundError('Gagal memperbarui catatan. Id tidak ditemukan');
     }
 
-    return result.rows.map(mapDBToModel)[0].id;
+    const { owner, ...updatedNote } = result.rows.map(mapDBToModel)[0];
+    await this._cacheService.delete(`notes:${owner}`);
+
+    return updatedNote.id;
   }
 
   /**
@@ -119,7 +135,7 @@ class NoteService {
    */
   async deleteNoteById(id) {
     const query = {
-      text: 'DELETE FROM notes WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM notes WHERE id = $1 RETURNING id, owner',
       values: [id],
     };
 
@@ -128,6 +144,9 @@ class NoteService {
     if (!result.rows.length) {
       throw new NotFoundError('Catatan gagal dihapus. Id tidak ditemukan');
     }
+
+    const { owner } = result.rows[0];
+    await this._cacheService.delete(`notes:${owner}`);
   }
 
   async verifyNoteOwner(id, owner) {
@@ -137,11 +156,11 @@ class NoteService {
     };
 
     const result = await this._pool.query(query);
-    
+
     if (!result.rows.length) {
       throw new NotFoundError('Catatan tidak ditemukan');
     }
-    
+
     const note = result.rows[0];
     if (note.owner !== owner) {
       throw new AuthorizationError('Anda tidak berhak mengakses resource ini');
